Added working header.
Replaced placeholder in layout with functioning header. It switches back
to the home page. The layoutController registers a method call in the
headerController and this way the homeButton can switch the center
content on click. We can get the instance of the headerController by
having an fx:id and then simply having an field variable with @FXML.