Added header.css, made center content dynamic and added Enum for paths
Removed css usage in app, added css to header view
Created FilePathEnum to centralize paths (single source of truth).
Added homepage.fxml which later is supposed to have the table for
orders.
Made the layout controller inject the contentSwitch consumer into all
its "SubViewControllers".
Created for this the SubViewRequestHandler (interface) , which lets the
layout able to not care which exact controller it is dealing with.
To give control to the current subView we created the throwingConsumer
(interface), so that the subView can handle any exceptions.